Move stream-related structs and Connection trait to xmpp::stream module
